Move example sketch to appropriately named folder
The Arduino IDE requires the sketch folder name to match the filename of the primary sketch file. On a filename case-sensitive OS like Linux, this includes matching case. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > LIBRARYNAME menu after the library is installed.